requirements.txt
annotated-types==0.7.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
docstring_parser==0.17.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.0
grpcio-status==1.71.2
idna==3.10
numpy==2.3.3
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
vertexai==1.71.1
wheel==0.45.1

# Runtime for server and container
Flask==3.0.3
google-genai==1.38.0
gunicorn==22.0.0
python-dotenv==1.0.1
